Move sentry to separate parent command
