+++ src/core.ts
@@ -12,7 +12,7 @@ import gitRelease from './release/program-git-release.js'
 import githubRelease from './release/program-github-release.js'
 import githubReleaseAssets from './release/program-github-release-assets.js'
 import githubPromoteRelease from './release/program-github-promote-release.js'
-import sentryRelease from './release/program-sentry-release.js'
+import sentryRelease from './sentry/program-sentry-release.js'
 import notify from './notify/notify.js'
 import packageJson from '../package.json' with { type: 'json' }
 
@@ -45,7 +45,9 @@ const buildCommand = (exitOverride?: (err: CommanderError) => never | void) => {
   githubRelease(releaseCommand)
   githubReleaseAssets(releaseCommand)
   githubPromoteRelease(releaseCommand)
-  sentryRelease(releaseCommand)
+
+  const sentryCommand = v0.command('sentry')
+  sentryRelease(sentryCommand)
 
   const notifyCommand = v0.command('notify')
   notify(notifyCommand)
